requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.4
scikit-learn>=1.3.0
scikit-surprise>=1.1.3
vaderSentiment>=3.3.2
plotly>=5.18.0
joblib>=1.3.0
tqdm>=4.66.0
streamlit-option-menu>=0.3.6
Pillow>=10.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
httpx>=0.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
